+++ src/quocslib/gradientfreemethods/DirectSearchMethod.py
