+++ src/quocslib/gradientfreemethods/NelderMead.py
@@ -17,18 +17,16 @@ import numpy as np
 
 np.seterr(all="raise")
 
-from quocslib.freegradientmethods.DirectSearchMethod import DirectSearchMethod
+from quocslib.gradientfreemethods.DirectSearchMethod import DirectSearchMethod
 from quocslib.stoppingcriteria.NelderMeadStoppingCriteria import (
     NelderMeadStoppingCriteria,
 )
 
 
-class FreeGradientTemplate(DirectSearchMethod):
+class NelderMead(DirectSearchMethod):
     callback: callable
 
-    def __init__(
-        self, settings: dict, stopping_criteria: dict, callback: callable = None
-    ):
+    def __init__(self, settings: dict, stopping_criteria: dict, callback: callable = None):
         """
         :param dict settings:
         :param dict stopping_criteria:
@@ -43,12 +41,10 @@ class FreeGradientTemplate(DirectSearchMethod):
         # Stopping criteria object
         self.sc_obj = NelderMeadStoppingCriteria(stopping_criteria)
 
-    def run_dsm(
-        self, func, x0, args=(), initial_simplex=None, max_iterations_number=None
-    ) -> dict:
+    def run_dsm(self, func, x0, args=(), initial_simplex=None, max_iterations_number=None, **kwargs) -> dict:
         """
 
-        :param callable func: Function tbe called at every function evaluation
+        :param callable func: Function to be called at every function evaluation
         :param np.array x0: initial point
         :param tuple args: Further arguments
         :param np.array initial_simplex: Starting simplex for the Nelder Mead evaluation
@@ -64,12 +60,6 @@ class FreeGradientTemplate(DirectSearchMethod):
             self.sc_obj.max_iterations_number = max_iterations_number
         # Initialize the iteration number
         iterations = 0
-        # Initialize hyper-parameters if any
-
-        # Other operations ...
-
-        # Initialize the function evaluation
-
         # Landscape dimension
         dim = len(x0)
         # Hyper-parameters for adaptive and not adaptive NM
