+++ src/quocslib/optimizationalgorithms/ADAlgorithm.py
