+++ src/quocslib/optimizationalgorithms/dCRABAlgorithm.py
@@ -17,7 +17,7 @@ import numpy as np
 
 from quocslib.Optimizer import Optimizer
 from quocslib.Controls import Controls
-from quocslib.freegradientmethods.NelderMead import NelderMead
+from quocslib.gradientfreemethods.NelderMead import NelderMead
 from quocslib.tools.linearalgebra import simplex_creation
 from quocslib.utils.dynamicimport import dynamic_import
