+++ src/quocslib/optimizationalgorithms/dCRABNoisyAlgorithm.py
@@ -18,7 +18,7 @@ from scipy.stats import norm
 
 from quocslib.Optimizer import Optimizer
 from quocslib.Controls import Controls
-from quocslib.freegradientmethods.NelderMead import NelderMead
+from quocslib.gradientfreemethods.NelderMead import NelderMead
 from quocslib.tools.linearalgebra import simplex_creation
 from quocslib.tools.randomgenerator import RandomNumberGenerator
